cleanup styles in top nav - reorganize nav components - remove unnecessary styles
